clean up unused files; change path to posting to dummy db to (hopefully) encompass all local machines
