+++ pages/api/index.js
@@ -1,24 +1,3 @@
-const cars = {
-  cars: {
-    Nissan: [
-      { model: "Sentra", doors: 4 },
-      { model: "Maxima", doors: 4 },
-      { model: "Skyline", doors: 2 },
-    ],
-    Ford: [
-      { model: "Taurus", doors: 4 },
-      { model: "Escort", doors: 4 },
-    ],
-
-    Toyota: [
-      { model: "Highlander", doors: 2 },
-      { model: "Rav4", doors: 1 },
-    ],
-  },
-};
-
-import { apiHandler, omit } from "../../helpers/api";
-
 export default apiHandler({
   get: getUsers,
 });
