+++ pages/data/users.json
@@ -34,5 +34,14 @@
         "id": 4,
         "dateCreated": "2022-05-11T21:18:01.533Z",
         "dateUpdated": "2022-05-11T21:18:01.533Z"
+    },
+    {
+        "firstName": "Halsey",
+        "lastName": "Quinn",
+        "username": "halsey",
+        "hash": "$2a$10$3YsDPCfq3s9yiUmNL0kWteEljTOsOo5kvSmXW4QylqI/.hm.cLg6i",
+        "id": 5,
+        "dateCreated": "2022-05-11T22:33:14.857Z",
+        "dateUpdated": "2022-05-11T22:33:14.857Z"
     }
 ]
